Send request from wear device to phone on startup
On phone the dataservice is triggered every 3 hours and
most of the times there is no change in the data. Hence, wear
doesn't recieve any updates about the data. If the data is not stored
on the wear device while the very first sync is made, it'll never recieve the
data unless it is updated. Hence adding a simple timstamp to the data. This
has its own downsides but atleast the wear device recieves the data.